% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_satellite_data.R
\name{download_satellite_data}
\alias{download_satellite_data}
\title{Télécharger une image satellite Sentinel-2}
\usage{
download_satellite_data(
  aoi,
  date,
  date_end = NULL,
  cloud_cover = 20,
  output_dir = "data/satellite",
  ...
)
}
\arguments{
\item{aoi}{sf ou bbox vecteur. Zone d'intérêt.}

\item{date}{Date de début (ex: "2023-06-01")}

\item{date_end}{Date de fin (par défaut = date + 1 mois)}

\item{cloud_cover}{Maximum de couverture nuageuse (0-100)}

\item{output_dir}{Dossier de sortie}

\item{...}{Autres paramètres pour \code{CDSE::get_image}}
}
\value{
Chemin du fichier raster téléchargé (ou un SpatRaster)
}
\description{
Télécharger une image satellite Sentinel-2
}
\examples{
\dontrun{
\dontrun{
  library(sf)
  aoi <- st_read("my_area.shp")
  download_satellite_data(aoi, "2023-06-01")
}
}
}
