% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_biodiversity}
\alias{sample_biodiversity}
\title{Données de biodiversité terrain simulées}
\format{
Un data.frame avec 200 lignes et 5 variables :
\describe{
\item{espece}{Nom de l'espèce observée}
\item{abondance}{Nombre d'individus comptés}
\item{parcelle}{Identifiant de la parcelle}
\item{lon}{Longitude}
\item{lat}{Latitude}
}
}
\description{
Données de biodiversité terrain simulées
}
\keyword{datasets}
