% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_landcover}
\alias{sample_landcover}
\title{Raster d'occupation du sol d'exemple}
\format{
Une matrice 20x20.
}
\source{
Donnees simulees
}
\usage{
sample_landcover
}
\description{
Une matrice simulee avec 5 classes d'occupation du sol.
}
\keyword{datasets}
