% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance_to_habitat.R
\name{calculate_distance_to_habitat}
\alias{calculate_distance_to_habitat}
\title{Calculer les distances aux habitats naturels}
\usage{
calculate_distance_to_habitat(
  landcover,
  points = NULL,
  habitat_codes = c(2, 3)
)
}
\arguments{
\item{landcover}{Objet \code{SpatRaster} d'occupation du sol.}

\item{points}{Objet sf avec les points d'observation (parcelles).
Si NULL, retourne uniquement les rasters de distance.}

\item{habitat_codes}{Vecteur des codes de classes d'habitat cibles.
Defaut: c(2, 3) = Foret + Prairie.}
}
\value{
Une liste contenant :
\describe{
\item{distance_raster}{SpatRaster de distance (en metres) aux habitats}
\item{points_with_distance}{Objet sf des points avec colonne dist_habitat ajoutee}
}
}
\description{
Calcule des rasters de distance euclidienne aux habitats d'interet
(foret, haies, habitats semi-naturels) et extrait ces distances
pour chaque parcelle d'observation.
}
\examples{
\dontrun{
\dontrun{
lc <- import_landcover()
bd <- import_biodiversity_data()
result <- calculate_distance_to_habitat(lc, bd$sf_object)
terra::plot(result$distance_raster, main = "Distance aux habitats naturels")
}
}
}
