% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_biodiversity}
\alias{sample_biodiversity}
\title{Donnees de biodiversite d'exemple}
\format{
Une liste avec plusieurs composantes.
}
\source{
Donnees simulees
}
\usage{
sample_biodiversity
}
\description{
Jeu de donnees simule pour les exemples du package.
}
\keyword{datasets}
